chatbot/app.py
from langchain_openai import ChatOpenAI
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

import streamlit as st
import os
from dotenv import load_dotenv

load_dotenv()

os.environ["OPENAI_API_KEY"]=os.getenv("OPENAI_API_KEY")
# langsmith tracking
os.environ["LANGCHAIN_TRACING_V2"]="true"
os.environ["LANGCHAIN_API_KEY"]=os.getenv("LANGCHAIN_API_KEY")

# prompt Template 

promt=ChatPromptTemplate.from_messages(
    [
        ("system", "You are a helpful assistant. Please response to the user question"),
        ("user", "Question:{question}")
    ]
)

# streamlit framework

st.title('Langchain Demo with OPENAI API')
input_text=st.text_input("Search the topic you want")

# openAI llm
llm=ChatOpenAI(model="gpt-3.5-turbo")
output_parser=StrOutputParser()
chain=promt|llm|output_parser

if input_text:
    st.write(chain.invoke({'question':input_text}))

print("OKE")