requirements.txt
langchain-openai
langchain-core
streamlit
python-dotenv